renames jobs to job posts
