+++ app/Http/Controllers/JobPostsController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-class JobsController extends Controller
+class JobPostsController extends Controller
 {
     //
 }
