+++ database/factories/JobPostsFactory.php
@@ -2,17 +2,17 @@
 
 namespace Database\Factories;
 
-use App\Models\Jobs;
+use App\Models\JobPost;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class JobsFactory extends Factory
+class JobPostsFactory extends Factory
 {
     /**
      * The name of the factory's corresponding model.
      *
      * @var string
      */
-    protected $model = Jobs::class;
+    protected $model = JobPost::class;
 
     /**
      * Define the model's default state.
